fix) controller 충돌 문제 해결
