+++ src/main/java/com/lionheart15/ideamarket/controller/UserController.java
@@ -1,12 +1,18 @@
 package com.lionheart15.ideamarket.controller;
 
+import com.lionheart15.ideamarket.domain.entity.dto.UserSignUpDto;
+import com.lionheart15.ideamarket.repository.UserRepository;
+import com.lionheart15.ideamarket.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
@@ -17,6 +23,10 @@ import javax.servlet.http.HttpServletResponse;
 @RequestMapping("/users")
 @Slf4j
 public class UserController {
+    @Autowired
+    private UserRepository userRepository;
+
+    private final UserService userService;
 
     @GetMapping("/login")
     public String loginPage() {
@@ -31,6 +41,18 @@ public class UserController {
             new SecurityContextLogoutHandler().logout(request, response, auth);
         }
         return "redirect:/";
+    }
+    @PostMapping("/create")
+    public  String create(@ModelAttribute UserSignUpDto dto ){
+
+        if(userRepository.existsByLoginId(dto.getLoginId())){
+
+            return "iddup";
+        }
+
+        userService.save(dto.getBirth(), dto.getEmail(), dto.getGender(),dto.getLoginId(),dto.getName(),dto.getPassword(),dto.getPhoneNumber());
+
+        return"login";
     }
 
     @GetMapping("/signup")
