+++ src/main/resources/templates/iddup.mustache
@@ -17,7 +17,7 @@
         </script>
 <br>
 <div>
-    <form  action="/create" method="post" >
+    <form  action="/users/create" method="post" >
       <h1>회원 가입</h1>
 
       <input type="text" name="name" placeholder="이름" "required/>
