+++ src/main/resources/templates/signup.mustache
@@ -16,7 +16,7 @@
         </script>
 <br>
 <div>
-    <form  action="/create" method="post" >
+    <form  action="/users/create" method="post" >
       <h1>회원 가입</h1>
 
       <input type="text" name="name" placeholder="이름" "required/>
